requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=hd4dd3e8_0
accelerate=0.16.0=pyhd8ed1ab_0
aiohttp=3.8.3=py37h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
arrow-cpp=8.0.0=py37h60b952e_0
async-timeout=4.0.2=py37h06a4308_0
asynctest=0.13.0=py_0
attrs=22.1.0=py37h06a4308_0
aws-c-common=0.4.57=he6710b0_1
aws-c-event-stream=0.1.6=h2531618_5
aws-checksums=0.1.9=he6710b0_0
aws-sdk-cpp=1.8.185=hce553d0_0
blas=1.0=mkl
boost-cpp=1.73.0=h7f8727e_12
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.0=h5eee18b_0
ca-certificates=2023.5.7=hbcca054_0
certifi=2023.5.7=pyhd8ed1ab_0
cffi=1.15.1=py37h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py37h06a4308_0
cryptography=39.0.1=py37h9ce1e76_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cudatoolkit=11.6.0=hecad31d_10
datasets=2.6.1=py37h06a4308_0
dill=0.3.6=py37h06a4308_0
einops=0.6.1=pyhd8ed1ab_0
feather-format=0.4.1=pyh9f0ad1d_0
ffmpeg=4.2.2=h20bf706_0
filelock=3.9.0=py37h06a4308_0
flit-core=3.6.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=py37h5eee18b_0
fsspec=2022.11.0=py37h06a4308_0
gflags=2.2.2=he6710b0_0
giflib=5.2.1=h5eee18b_3
glog=0.5.0=h2531618_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
grpc-cpp=1.46.1=h33aed49_1
huggingface_hub=0.10.1=py37h06a4308_0
icu=58.2=he6710b0_3
idna=3.4=py37h06a4308_0
importlib-metadata=4.11.3=py37h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.1.1=py37h06a4308_0
jpeg=9e=h5eee18b_1
krb5=1.19.4=h568e23c_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libboost=1.73.0=h28710b8_12
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.6.1.9=0
libcurand=10.3.2.106=0
libcurl=7.88.1=h91b91d3_0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_0
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=h8f2d780_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libidn2=2.3.4=h5eee18b_0
libnghttp2=1.46.0=hce63b2e_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libopus=1.3.1=h7b6447c_0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libssh2=1.10.0=h8f2d780_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libthrift=0.15.0=hcc01f38_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lz4-c=1.9.4=h6a678d5_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
multidict=6.0.2=py37h5eee18b_0
multiprocess=0.70.14=py37h06a4308_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
nltk=3.7=pyhd3eb1b0_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h7f8727e_0
orc=1.7.4=hb3bc3d3_1
packaging=22.0=py37h06a4308_0
pandas=0.23.4=py37hf8a1672_0
pillow=9.4.0=py37h6a678d5_0
pip=22.3.1=py37h06a4308_0
psutil=5.9.0=py37h5eee18b_0
pyarrow=8.0.0=py37h992f0b0_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.0.0=py37h06a4308_0
pysocks=1.7.1=py37_1
python=3.7.15=h7a1cb2a_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-xxhash=2.0.2=py37h5eee18b_1
python_abi=3.7=2_cp37m
pytorch=1.12.1=py3.7_cuda11.6_cudnn8.3.2_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytz=2022.7=py37h06a4308_0
pyyaml=6.0=py37h5eee18b_1
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
regex=2022.7.9=py37h5eee18b_0
requests=2.28.1=py37h06a4308_0
responses=0.13.3=pyhd3eb1b0_0
setuptools=65.6.3=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h295c915_0
sqlite=3.41.2=h5eee18b_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.12.1=py37_cu116
torchvision=0.13.1=py37_cu116
tqdm=4.64.1=py37h06a4308_0
typing-extensions=4.4.0=py37h06a4308_0
typing_extensions=4.4.0=py37h06a4308_0
urllib3=1.26.14=py37h06a4308_0
utf8proc=2.6.1=h27cfd23_0
wheel=0.38.4=py37h06a4308_0
x264=1!157.20191217=h7b6447c_0
xxhash=0.8.0=h7f8727e_3
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.8.1=py37h5eee18b_0
zipp=3.11.0=py37h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
